+++ ffc/build.gradle
@@ -48,7 +48,7 @@ dependencies {
   implementation "com.squareup.retrofit2:retrofit:2.4.0"
   implementation "com.squareup.retrofit2:converter-gson:2.4.0"
   implementation "org.jetbrains.kotlin:kotlin-stdlib-jre7:$kotlin_version"
-  implementation "org.jetbrains.anko:anko:0.10.4"
+  implementation "org.jetbrains.anko:anko-appcompat-v7-commons:0.10.4"
   implementation "me.piruin:spinney:0.2.2"
   implementation "com.google.code.gson:gson:2.8.2"
