+++ ffc/src/main/java/ffc/v3/api/AuthTokenInterceptor.kt
@@ -15,33 +15,20 @@
  * limitations under the License.
  */
 
-package ffc.v3
+package ffc.v3.api
 
-import android.os.Build
 import okhttp3.Interceptor
 import okhttp3.Response
 import java.io.IOException
-import java.lang.String.format
 
-internal class DefaultInterceptor : Interceptor {
+class AuthTokenInterceptor(private val token: String) : Interceptor {
 
   @Throws(IOException::class)
   override fun intercept(chain: Interceptor.Chain): Response {
     val original = chain.request()
-
     val builder = original.newBuilder()
-    builder.addHeader("User-Agent", USER_AGENT)
-      .addHeader("Accept", "application/json; charset=utf-8")
-      .addHeader("Accept-Charset", "utf-8")
-      .addHeader("X-Requested-By", "ffc-v3")
-    builder.method(original.method(), original.body())
+    builder.addHeader("Authorization", "Bearer $token")
 
     return chain.proceed(builder.build())
-  }
-
-  companion object {
-    private val ANDROID = format("Android ${Build.VERSION.RELEASE} (${Build.VERSION.SDK_INT})")
-    private val DEVICE = format("%s %s", Build.BRAND, Build.MODEL)
-    private val USER_AGENT = "FFC/${BuildConfig.VERSION_NAME} ($ANDROID; $DEVICE)"
   }
 }
