+++ ffc/src/main/java/ffc/v3/util/View.kt
@@ -15,23 +15,14 @@
  * limitations under the License.
  */
 
-package ffc.v3
+package ffc.v3.util
 
-import retrofit2.Call
-import retrofit2.http.GET
-import retrofit2.http.Header
-import retrofit2.http.POST
-import retrofit2.http.Path
+import android.view.View
 
-interface OrgService {
+fun View.gone() {
+  this.visibility = View.GONE
+}
 
-  @GET("org") fun listOrgs(): Call<List<Org>>
-
-  @GET("org?my=true") fun myOrg(): Call<List<Org>>
-
-  @POST("org/{id}/authorize")
-  fun createAuthorize(
-    @Path("id") id: Long,
-    @Header("Authorization") authorize: String
-  ): Call<Authorize>
+fun View.visible() {
+  this.visibility = View.VISIBLE
 }
